+++ tests/dummy/app/routes/signin.js
@@ -2,24 +2,17 @@ import Ember from 'ember';
 import ENV from '../config/environment';
 export default Ember.Route.extend({
 
-  //
-  // onDeactivate: function(){
-  //   Ember.debug('route:signin:deactivate fired...');
-  // }.on('deactivate'),
 
   deactivate: function(){
+    //if you are using iframes, you will need to remove the
+    //iframe from the DOM here so torii gets notified that the
+    //auth attempt was cancelled.
     Ember.debug('route:signin:deactivate fired...');
   },
 
   actions: {
     signin: function(){
-
-      let parent = window.location.protocol + '//' + window.location.hostname;
-
-      this.get('session').open('arcgis-oauth-bearer', {
-          parent: parent,
-          display:'iframe'
-        })
+      this.get('session').open('arcgis-oauth-bearer')
         .then((authorization) => {
           Ember.debug('AUTH SUCCESS: ', authorization);
           //transition to some secured route or... so whatever is needed
@@ -30,31 +23,30 @@ export default Ember.Route.extend({
         });
     },
 
+
     /**
      * In order to show the iframe when this route loads
      * we add a didTransition hook, and then schedule the
      * session.open to be called after it's rendered
      */
     didTransition: function(){
-      Ember.debug('route:signin:didTransition');
+      //only do this if we are using iframe style
+      if(ENV.torii.providers['arcgis-oauth-bearer'].display && ENV.torii.providers['arcgis-oauth-bearer'].display === 'iframe'){
+        //--- USE THIS BLOCK IN YOUR APP --
+        Ember.run.schedule('afterRender', this, function(){
+          this.get('session').open('arcgis-oauth-bearer')
+            .then((authorization) => {
+              Ember.debug('AUTH SUCCESS: ', authorization);
+              //transition to secured route
+              this.controller.transitionToRoute('secure');
+            })
+            .catch((err)=>{
+              Ember.debug('AUTH ERROR: ' + JSON.stringify(err));
+            });
+        });
+        //-----------------------------------
+      }
 
-      this.controller.set('showSignInButton', false);
-
-      let parent = window.location.protocol + '//' + window.location.hostname;
-      Ember.run.schedule('afterRender', this, function(){
-        this.get('session').open('arcgis-oauth-bearer', {
-            parent:parent,
-            display:'iframe'
-          })
-          .then((authorization) => {
-            Ember.debug('AUTH SUCCESS: ', authorization);
-            //transition to secured route
-            this.controller.transitionToRoute('secure');
-          })
-          .catch((err)=>{
-            Ember.debug('AUTH ERROR: ' + JSON.stringify(err));
-          });
-      });
     }
   }
 });