+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.0.1]
 ### Added
+- support for iframe (\*.argis.com only apps) oAuth
+- support for application (pop-up) oAuth